+++ ImageServer/Web/Common/Exceptions/ExceptionHandler.cs
@@ -26,6 +26,7 @@ using System;
 using System.Web;
 using ClearCanvas.Common;
 using System.Text;
+using ClearCanvas.Common.Rest;
 using ClearCanvas.ImageServer.Web.Common.Security;
 
 namespace ClearCanvas.ImageServer.Web.Common.Exceptions
@@ -52,17 +53,36 @@ namespace ClearCanvas.ImageServer.Web.Common.Exceptions
 		{
 			context = HttpContext.Current;
 			Platform.Log(LogLevel.Error, e);
+			
+			// Handle cross-site script attacks.
+			// They can cause HttpRequestValidationException or HttpException with a special error message.
+			// They can also generate HttpException without any message. We  should handle all of them.
+			// We should also avoid using context.Server.Transfer() because it will end up throwing exception too.
+			// Instead, we should call Response.RedirectPermanent
+			var sb = new StringBuilder();
+			if (e is HttpRequestValidationException || (e is HttpException &&
+				(string.IsNullOrEmpty(e.Message) || e.Message.StartsWith("A potentially dangerous"))))
+			{
+				HttpContext.Current.Response.RedirectPermanent(ImageServerConstants.PageURLs.InvalidRequestErrorPage, true);
+				return;
+			}
+
+
 			if (context.Items.Contains(ImageServerConstants.ContextKeys.ErrorMessage))
 				context.Items.Remove(ImageServerConstants.ContextKeys.ErrorMessage);
 			if (context.Items.Contains(ImageServerConstants.ContextKeys.StackTrace))
 				context.Items.Remove(ImageServerConstants.ContextKeys.StackTrace);
 
-            StringBuilder sb = new StringBuilder();
-            sb.AppendLine(e.Message);
-
 			context.Items.Add(ImageServerConstants.ContextKeys.ErrorMessage, sb.ToString());
 			context.Items.Add(ImageServerConstants.ContextKeys.StackTrace, e.StackTrace);
-			context.Server.Transfer(ImageServerConstants.PageURLs.ErrorPage);
+			try
+			{
+				context.Server.Transfer(ImageServerConstants.PageURLs.ErrorPage);
+			}
+			catch (Exception ex)
+			{
+				Platform.Log(LogLevel.Error, ex);
+			}
 		}
 
     	public static void ThrowException(BaseWebException e)
